requirements.txt
azure-functions
azure-storage-blob
azure-search-documents
openai
pymupdf
langchain-text-splitters
pydantic-settings